Protect user routes and update README

+++ user_service/app/routers/user.py
@@ -1,11 +1,14 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from user_service.app.models.user_models import User as UserModel
 from user_service.app.schemas.user import (
     UserCreate,
     UserRead,
     UserUpdate
     )
+
+from user_service.app.auth.dependencies import get_current_user, get_current_admin
 from user_service.app.core.db_depends import get_async_db
 from user_service.app.services.service import (
     create_user,
@@ -34,13 +37,13 @@ async def create_user_route(
 
 
 @router.get("/active", response_model=list[UserRead], status_code=status.HTTP_200_OK)
-async def get_list_active_users_route(db : AsyncSession = Depends(get_async_db)):
+async def get_list_active_users_route( _: UserModel = Depends(get_current_admin), db : AsyncSession = Depends(get_async_db)):
     users = await get_list_active_users(db)
     return users
 
 
 @router.get("/inactive", response_model=list[UserRead], status_code=status.HTTP_200_OK)
-async def get_list_inactive_users_route(db : AsyncSession = Depends(get_async_db)):
+async def get_list_inactive_users_route( _: UserModel = Depends(get_current_admin), db : AsyncSession = Depends(get_async_db)):
     users = await get_list_inactive_users(db)
     return users
 
@@ -48,8 +51,11 @@ async def get_list_inactive_users_route(db : AsyncSession = Depends(get_async_db
 @router.get("/{user_id}", response_model=UserRead, status_code=status.HTTP_200_OK)
 async def get_user_by_id_route(
     user_id : int ,
+    current_user : UserModel = Depends(get_current_user),
     db : AsyncSession = Depends(get_async_db)
     ):
+    if current_user.id != user_id and current_user.role.value != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
     user = await get_user_by_id(db, user_id)
     
     if user is None:
@@ -63,7 +69,10 @@ async def get_user_by_id_route(
 
 
 @router.patch("/{user_id}", response_model=UserRead, status_code=status.HTTP_200_OK)
-async def update_user_route(user_id : int, user_data : UserUpdate, db : AsyncSession = Depends(get_async_db)):
+async def update_user_route(user_id : int, user_data : UserUpdate, current_user : UserModel = Depends(get_current_user), db : AsyncSession = Depends(get_async_db)):
+    if current_user.id != user_id and current_user.role.value != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
+    
     try:
         user = await update_user(db, user_id, user_data )
     except ValueError as exc:
@@ -76,7 +85,13 @@ async def update_user_route(user_id : int, user_data : UserUpdate, db : AsyncSes
 
 
 @router.delete("/{user_id}", response_model=UserRead, status_code=status.HTTP_200_OK)
-async def delete_user_route(user_id : int, db : AsyncSession = Depends(get_async_db)):
+async def delete_user_route(
+    user_id : int, 
+    current_user: UserModel = Depends(get_current_user),
+    db : AsyncSession = Depends(get_async_db)):
+    if current_user.id != user_id and current_user.role.value != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
+    
     try:
         user = await delete_user(db, user_id)
     except ValueError as exc:
